cloud-function/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==2.0.3
fsspec==2023.6.0
gcsfs==2023.6.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.27.1
google-cloud-aiplatform[pipelines]